test: Refactor x86_64 test suite structure and add comprehensive arithmetic coverage
This commit reorganizes the x86_64 integration tests into a unified crate structure, improves the test harness, and significantly expands coverage for arithmetic instructions.

- tests/x86_64.rs (modified):
  - Update module declarations to include all test files with a consistent naming scheme.
  - Mark as auto-generated to reflect the aggregated nature of the suite.

- tests/x86_64/common/mod.rs (modified):
  - Introduce `VM` struct to encapsulate VCPU state and memory for tests.
  - Add `setup_vm_legacy` and `run_until_hlt_legacy` helpers to support existing test patterns.
  - Add `TestCase` struct for hex-string based tests.

- tests/x86_64/arithmetic/comprehensive_arithmetic.rs (added):
  - Add extensive tests for Group 1 arithmetic instructions: `adc`, `add`, `sub`, `sbb`, `inc`, `dec`, `neg`, `mul`, `div`, `idiv`.
  - Cover edge cases including carry chains, overflows, zero flags, and sign extension.

- tests/x86_64/**/*.rs (modified):
  - Replace relative `#[path = "../common/mod.rs"] mod common;` imports with `use crate::common::*;` to utilize the shared common module.
  - Update I/O handling to use `vcpu.complete_io_in()` instead of `vcpu.set_io_result()`.
  - Fix instruction encodings in `adc_extended.rs`, `sbb_extended.rs`, and `sub_extended.rs`.
  - Standardize memory initialization in SIMD tests to use explicit byte arrays rather than `vec!` repetition for clarity.

Impact:
- Improves test suite maintainability by centralizing common logic and standardizing imports.
- Increases reliability of arithmetic instruction implementation through rigorous edge-case testing.
- Aligns test harness with the current VCPU I/O API.

+++ tests/x86_64/rotate/rol_ror_extended.rs
@@ -485,6 +485,6 @@ fn test_rol_power_of_two() {
         let regs = run_until_hlt(&mut vcpu).unwrap();
 
         let expected = ((value << 3) | (value >> 5)) & 0xFF;
-        assert_eq!(regs.rax & 0xFF, expected, "ROL power of 2 value 0x{:02X}", value);
+        assert_eq!(regs.rax & 0xFF, expected as u64, "ROL power of 2 value 0x{:02X}", value);
     }
 }
